Split out http conversions in new module
